+++ test/fixtures/middlewares/middleware2.js
@@ -1,0 +1,3 @@
+export default function ({route, store}) {
+    console.log('middleware2 is running...');
+}