+++ test/integration/common.js
@@ -20,14 +20,11 @@ test.beforeEach('init lavas-core & server', async t => {
 });
 
 test.afterEach('clean', t => {
-    server.close();
+    server && server.close();
 });
 
-test('it should run in development mode correctly', async t => {
-
-    // init, build and start a server
-    await core.init('development', true);
-    await core.build();
+async function runCommonTestCases(t) {
+    // set middlewares & start a server
     app.use(isKoaSupport ? core.koaMiddleware() : core.expressMiddleware());
     server = app.listen(port, () => {
         console.log('server started at localhost:' + port);
@@ -37,7 +34,7 @@ test('it should run in development mode correctly', async t => {
     let ssrContent = '<div id="app" data-server-rendered="true">';
     let indexPageTitle = '<title data-vue-meta="true">Home</title>';
     let res = await request(app)
-        .get('/');
+        .get('/?a=1&b=2');
     t.is(200, res.status);
     t.true(res.text.indexOf(ssrContent) > -1);
     t.true(res.text.indexOf(indexPageTitle) > -1);
@@ -51,11 +48,10 @@ test('it should run in development mode correctly', async t => {
     t.true(res.text.indexOf(childPageTitle) > -1);
 
     // serve static assets such as manifest.json
-    let manifestContent = core.config.manifest;
     res = await request(app)
         .get('/static/manifest.json');
     t.is(200, res.status);
-    t.is(res.text, JSON.stringify(manifestContent));
+    t.true(res.text.startsWith(`{"start_url":"/?utm_source=homescreen",`));
 
     // render error page for invalid route path
     let errorPageTitle = '<title data-vue-meta="true">服务器开小差了</title>';
@@ -63,4 +59,25 @@ test('it should run in development mode correctly', async t => {
         .get('/invalid/path');
     t.is(200, res.status);
     t.true(res.text.indexOf(errorPageTitle) > -1);
+}
+
+test.serial('it should run in development mode correctly', async t => {
+    // init, build and start a dev server
+    await core.init('development', true);
+    await core.build();
+
+    await runCommonTestCases(t);
+});
+
+test.serial('it should run in production mode correctly', async t => {
+    // build in production mode
+    await core.init('production', true);
+    await core.build();
+
+    // start server in production mode
+    core = new LavasCore(join(__dirname, '../fixtures/dist'));
+    await core.init('production');
+    await core.runAfterBuild();
+
+    await runCommonTestCases(t);
 });